[mlir] fix warning

+++ enzyme/Enzyme/MLIR/Dialect/Ops.h
@@ -48,7 +48,7 @@ template <typename SourceOp, bool filterGrad, bool includeShadows = true,
           bool includeDifferentialReturns = true>
 llvm::SmallVector<mlir::Value, 2> filterGradInputs(SourceOp uop) {
   llvm::SmallVector<mlir::Value, 2> outs;
-  size_t in_idx = 0;
+  unsigned in_idx = 0;
 
   for (auto act : uop.getActivity()) {
     auto iattr = cast<ActivityAttr>(act);
